Automation/Yts Discord Bot/requirements.txt
discord.py==1.7.1
requests==2.25.0